requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
fastapi==0.115.12
Flask==3.1.1
gevent==25.5.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jq==1.8.0
MarkupSafe==3.0.2
numpy==2.2.6
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
qrcode==8.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websocket==0.2.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
